Revamping Settings page to split into menu bar
