+++ packages/manager/src/views/Settings/components/AccessControl/AddMembers/index.tsx
@@ -1,0 +1,1 @@
+export { AddMembersForm } from './AddMembersForm';
