+++ packages/manager/src/views/Settings/components/AccessControl/index.tsx
@@ -1,0 +1,1 @@
+export { AccessControl } from './AccessControl';
