+++ packages/manager/src/views/Settings/components/ApiKey/index.tsx
@@ -1,0 +1,1 @@
+export { ApiKey } from './ApiKey';
