+++ packages/manager/src/views/Settings/components/CmdbDetails/index.tsx
@@ -1,0 +1,1 @@
+export { CmdbDetails } from './CmdbDetails';
