+++ packages/manager/src/views/Settings/components/DeleteWebProperty/index.tsx
@@ -1,0 +1,1 @@
+export { DeleteWebProperty } from './DeleteWebProperty';
