+++ packages/manager/src/views/Settings/components/Environment/index.tsx
@@ -1,0 +1,1 @@
+export { Environment } from './Environment';
