+++ packages/manager/src/views/Settings/index.tsx
@@ -1,0 +1,1 @@
+export { Settings } from './Settings';
