--- packages/manager/src/views/WebPropertyEnvPage/components/AddMembers/index.tsx
@@ -1,1 +1,0 @@
-export { AddMembers } from './AddMembers';
