--- packages/manager/src/views/WebPropertyEnvPage/index.tsx
@@ -1,1 +1,0 @@
-export { WebPropertyEnvPage } from './WebPropertyEnvPage';
